idempotency check feature

+++ routes/index.js
@@ -23,16 +23,32 @@ router.post('/uploadDocument', uploadController.upload);
 
 router.get('/download/:file', (req, res) => {
   const { file } = req.params;
+  if(file==undefined ||file ==''){
+    return res.json({success:false, message:"File name cannot be null/empty"})
+  }
   const fileLocation = path.join('./uploads', file);
   res.download(fileLocation, file);
 });
 
 router.post('/verify',(request, res)=>{
+  console.log(request.body)
+  if(request.body.verifierAddress==undefined||request.body.verifierAddress==''){
+    return res.status(400).json({success:false, message:"Verifier Address cannot be empty"})
+  }
+  if(request.body.userId==undefined||request.body.userId==''){
+    return res.status(400).json({success:false, message:"UserId cannot be empty"})
+  }
+  if(request.body.originalData==undefined||request.body.originalData==''){
+    return res.status(400).json({success:false, message:"Original Data cannot be empty"})
+  }
+  if(request.body.userPublicKey==undefined||request.body.userPublicKey==''){
+    return res.status(400).json({success:false, message:"User Public Key cannot be empty"})
+  }
   const kycData = new KycData({
-    verifierAddress: request.body.verifierAddress, userId: request.body.userId, data: request.body.originalData,
+    verifierAddress: request.body.verifierAddress, userId: request.body.userId, data: request.body.originalData, userPublicKey:request.body.userPublicKey
   });
   kycData.save((error, data) => {
-    if(error) res.status(500).json({success: false });
+    if(error) res.status(500).json({success: false, message:"Error saving to Database", error });
     else {
       res.status(200).json({success:true, message:"verified"})
     }
@@ -40,18 +56,24 @@ router.post('/verify',(request, res)=>{
 })
 
 router.post('/request/delete', (req, res) => {
+  if(req.body._id==undefined||req.body._id==''){
+    return res.status(400).json({success:false, message:"_id cannot be empty"})
+  }
   const { _id } = req.body;
   console.log(req.body._id);
   Request.deleteOne({ _id }, (error) => {
-    if (error) res.status(500).json({ success: false, error });
+    if (error) res.status(500).json({ success: false, message:"Error deleting from database",error });
     else res.status(200).json({ success: true });
   });
 });
 
 router.get('/kycData', (req, res) => {
   const { verifierAddress } = req.query;
+  if(verifierAddress==undefined||verifierAddress==''){
+    return res.status(400).json({success:false, message:"Verifier Address cannot be empty"})
+  }
   KycData.find({ verifierAddress }, (error, data) => {
-    if (error) res.status(500).json({ success: false, error });
+    if (error) res.status(500).json({ success: false, message:"Error finding data from database",error });
     else res.status(200).json({ success: true, data });
   });
 });
@@ -64,9 +86,18 @@ router.post('/mailQR', qrmailer.qr);
 router.post('/verifyOTP',(req,res)=>{
 
   const { _id, otp, originalData }= req.body;
+  if(_id==undefined||_id==''){
+    return res.status(400).json({success:false, message:"_id cannot be empty"})
+  }
+  if(otp==undefined||otp==''){
+    return res.status(400).json({success:false, message:"otp cannot be empty"})
+  }
+  if(originalData==undefined||originalData==''){
+    return res.status(400).json({success:false, message:"data cannot be empty"})
+  }
   VerificationRequest.find({ _id }, function (err,requests) {
     // console.log(requests)
-    if (err || requests.length==0) return res.json({success:false,message:"Could not locate reuqest"})
+    if (err || requests.length==0) return res.json({success:false,message:"Could not locate reuqest",err})
     var request = requests[0]
     if(request.otp == otp){ 
       console.log()
@@ -77,14 +108,14 @@ router.post('/verifyOTP',(req,res)=>{
       try{
         var verified=publicKey.verify(md.digest().bytes(),forge.util.decode64(request.signature))
       }catch(e){
-        return res.status(401).json({success:false,message:"Invalid Data"})
+        return res.status(401).json({success:false,message:"Mismatched Keys",e})
       }
       if (verified==true){
         const kycData = new KycData({
           verifierAddress: request.verifierAddress, userId: request.userId, data: originalData,
         });
         kycData.save((error, data) => {
-          if(error) res.status(500).json({success: false });
+          if(error) res.status(500).json({success: false ,message:"Error saving to Db"});
           else {
             VerificationRequest.findByIdAndDelete(_id, (error) => {
               if(error) res.status(500).json({ success: false });
@@ -108,12 +139,32 @@ router.post('/sendMail', emailController.email);
 
 router.post('/initiateVerification',(req, res)=>{
   const {otp, verifierAddress, userId, userPublicKey, verifierPublicKey, signature, email,_id, encryptedData} = req.body;
+  if(_id==undefined||_id==''){
+    return res.status(400).json({success:false, message:"_id cannot be empty"})
+  }
+  if(otp==undefined||otp==''){
+    return res.status(400).json({success:false, message:"otp cannot be empty"})
+  }if(verifierAddress==undefined||verifierAddress==''){
+    return res.status(400).json({success:false, message:"verifier address cannot be empty"})
+  }if(userId==undefined||userId==''){
+    return res.status(400).json({success:false, message:"user Id cannot be empty"})
+  }if(userPublicKey==undefined||userPublicKey==''){
+    return res.status(400).json({success:false, message:"user Public key cannot be empty"})
+  }if(verifierPublicKey==undefined||verifierPublicKey==''){
+    return res.status(400).json({success:false, message:"verifier public key annot be empty"})
+  }if(signature==undefined||signature==''){
+    return res.status(400).json({success:false, message:"signature cannot be empty"})
+  }if(email==undefined||email==''){
+    return res.status(400).json({success:false, message:"email cannot be empty"})
+  }
+  if(encryptedData==undefined||encryptedData==''){
+    return res.status(400).json({success:false, message:"encrypted data cannot be empty"})
+  }
   const newRequest = new VerificationRequest({
     verifierAddress, verifierPublicKey, userId, otp, signature, email
   });
   newRequest.save((error, request) => {
-    if (error) res.status(500).json({ success: false, error });
-    else res.status(200).json({ success: true, request });
+    if (error) return res.status(500).json({ success: false, error, message:"Error saving new request" });
   
   // console.log(userPublicKey)
   try{
@@ -141,8 +192,12 @@ router.post('/initiateVerification',(req, res)=>{
       console.log(bd)
       axios.post(url+'request/delete',bd)
     }
+    else{
+     return res.json({success:false,message:"Could not send email"});
+    }
   })
 });
+  return res.status(200).json({success:true, message:"OTP generated and mailed"})
 })
 
 module.exports = router;
